LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-options-credentials-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected string literal './credentials.py'. import call expects exactly one argument.

Harness Error (FAIL), message = SyntaxError: Unexpected string literal './credentials.py'. import call expects exactly one argument.

PASS Test initialization: setting up cross-origin cookie
PASS new Worker() with type=module and default credentials option should behave as credentials=same-origin and send the credentials
FAIL new Worker() with type=module and credentials=omit should not send the credentials assert_equals: expected "" but got "1"
PASS new Worker() with type=module and credentials=same-origin should send the credentials
PASS new Worker() with type=module and credentials=include should send the credentials
TIMEOUT new Worker() with type=module and default credentials option should behave as credentials=same-origin and send the credentials for same-origin static imports Test timed out
NOTRUN new Worker() with type=module and credentials=omit should not send the credentials for same-origin static imports
NOTRUN new Worker() with type=module and credentials=same-origin should send the credentials for same-origin static imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for same-origin static imports
NOTRUN new Worker() with type=module and default credentials option should behave as credentials=same-origin and not send the credentials for cross-origin static imports
NOTRUN new Worker() with type-module credentials=omit should not send the credentials for cross-origin static imports
NOTRUN new Worker() with type=module and credentials=same-origin should not send the credentials for cross-origin static imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for cross-origin static imports
NOTRUN new Worker() with type=module and default credentials option should behave as credentials=same-origin and send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=omit should not send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=same-origin should send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for same-origin dynamic imports
NOTRUN new Worker() with type=module and default credentials option should behave as credentials=same-origin and not send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type-module credentials=omit should not send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=same-origin should not send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type=module and credentials=include should send the credentials for cross-origin dynamic imports
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (default).
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (omit).
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (same-origin).
NOTRUN new Worker() with type=classic should always send the credentials regardless of the credentials option (include).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (default).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (omit).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (same-origin).
NOTRUN new Worker() with type=classic should always send the credentials for same-origin dynamic imports regardless of the credentials option (include).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (default).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (omit).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (same-origin).
NOTRUN new Worker() with type=classic should never send the credentials for cross-origin dynamic imports regardless of the credentials option (include).
